Added Actix, Reqwest and Tokio as dependencies for the email-newsletter project.
The commit includes the addition of Actix-web, in order to handle HTTP and other protocols in the email-newsletter project. Moreover, Reqwest was also added for handling HTTP Client tasks, and Tokio was included as a runtime for the project. Additionally, few Rust crates such as 'cargo' and 'cargo.lock' were also updated in this commit. The updated version may contain important fixes and improvements beneficial for the project.